sessions: refactor way attr subsys is setup
The win, datatype, and comm subsystems are dependent on
the attr subsys in a way that prevents a simple uncoordinated
tear down of the attr subsys.  Refactor the way attr subsys
is initialized to be based on when consumers of its functionality
are instantiated.

The handling of the predefined keyvals had to be moved out and
invoked separately as part of the creation of ompi_mpi_comm_world.

Fixes #18

Signed-off-by: Howard Pritchard <howardp@lanl.gov>